+++ examples/reader.rs
@@ -1,14 +1,17 @@
 use termal::{
+    codes,
     error::Result,
-    raw::{disable_raw_mode, enable_raw_mode, readers::prompt},
+    raw::{enable_raw_mode, readers::prompt},
+    reset_terminal,
 };
 
 fn main() -> Result<()> {
+    print!("{}", codes::ENABLE_BRACKETED_PASTE_MODE);
     enable_raw_mode()?;
 
     start()?;
 
-    disable_raw_mode()?;
+    reset_terminal();
 
     Ok(())
 }