+++ termal_core/src/raw/events/mod.rs
@@ -2,7 +2,8 @@ mod csi;
 mod event;
 mod key;
 pub mod mouse;
+mod state_change;
 mod status;
 mod term_attr;
 
-pub use self::{event::*, key::*, status::*, term_attr::*};
+pub use self::{event::*, key::*, state_change::*, status::*, term_attr::*};