+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 - Add `SS3` code.
 - Add code `OVERLINE`, also add to print macros as `overline`/`ol`.
 - Add codes to enable and disable inverse color in whole terminal.
+- Add option to enable bracketed paste mode (pasted text is verbatim).
+- Support bracketed paste mode in terminal.
 + Move around some logic around sixels.
 + Refactor readers.
 + Split into features.